requirements.txt
PyYAML==3.11
cffi==0.8.6
cryptography==0.6.1
pyOpenSSL==0.14
pycparser==2.10
python-etcd==0.3.2
six==1.8.0
urllib3==1.9.1
Jinja2==2.7.3